chore: remove blog post page component
