+++ apps/web/vibes/soul/examples/pages/blog/post/warm.tsx
@@ -1,17 +1,90 @@
-import { BlogPostPage } from '@/vibes/soul/pages/blog/post'
+import { posts } from '@/vibes/soul/examples/sections/blog-post-list'
+import {
+  contactInformation,
+  copyright,
+  footerLinks,
+} from '@/vibes/soul/examples/sections/footer/warm'
+import { headerLinks, logo } from '@/vibes/soul/examples/sections/header/warm'
+import { AnnouncementBar } from '@/vibes/soul/primitives/announcement-bar'
+import { BlogPostContent } from '@/vibes/soul/sections/blog-post-content'
+import { FeaturedBlogPostCarousel } from '@/vibes/soul/sections/featured-blog-post-carousel'
+import { Footer } from '@/vibes/soul/sections/footer'
+import {
+  Amex,
+  ApplePay,
+  Bitcoin,
+  GooglePay,
+  Mastercard,
+  Paypal,
+  Visa,
+} from '@/vibes/soul/sections/footer/payment-icons'
+import { Facebook, Instagram, X, Youtube } from '@/vibes/soul/sections/footer/social-icons'
+import { Header } from '@/vibes/soul/sections/header'
+import { Subscribe } from '@/vibes/soul/sections/subscribe'
+
+const socialMediaLinks = [
+  {
+    href: '#',
+    icon: <Facebook />,
+  },
+  {
+    href: '#',
+    icon: <X />,
+  },
+  {
+    href: '#',
+    icon: <Instagram />,
+  },
+  {
+    href: '#',
+    icon: <Youtube />,
+  },
+]
+
+const paymentIconsArray: React.ReactNode[] = [
+  <Visa key="Visa" />,
+  <Amex key="Amex" />,
+  <Mastercard key="Mastercard" />,
+  <Paypal key="Paypal" />,
+  <GooglePay key="GooglePay" />,
+  <ApplePay key="ApplePay" />,
+  <Bitcoin key="Bitcoin" />,
+]
+
+const locales = [
+  { id: '1', region: 'US', language: 'EN' },
+  { id: '2', region: 'FR', language: 'FR' },
+  { id: '3', region: 'DE', language: 'DC' },
+  { id: '4', region: 'IT', language: 'IT' },
+]
 
 export default function Preview() {
   return (
-    <BlogPostPage
-      id="1"
-      title="The Good, the Great, and the (My) Bad: Adventures of a First Time Bikepacker"
-      author="Sam Smith"
-      date="June 30, 2024"
-      image={{
-        src: 'https://rstr.in/monogram/vibes/oyasr_z-Tgb',
-        alt: 'A bicycle parked on a grassy cliff overlooking the ocean.',
-      }}
-      content={`
+    <>
+      <AnnouncementBar>
+        Get <strong>15% off</strong> and free shipping with discount code{' '}
+        <strong>&quot;welcome&quot;</strong>
+      </AnnouncementBar>
+
+      <Header
+        links={headerLinks}
+        logo={logo}
+        cartHref="#"
+        accountHref="#"
+        locales={locales}
+        activeLocale="EN"
+      />
+
+      <BlogPostContent
+        id="1"
+        title="The Good, the Great, and the (My) Bad: Adventures of a First Time Bikepacker"
+        author="Sam Smith"
+        date="June 30, 2024"
+        image={{
+          src: 'https://rstr.in/monogram/vibes/oyasr_z-Tgb',
+          alt: 'A bicycle parked on a grassy cliff overlooking the ocean.',
+        }}
+        content={`
         <p>A couple weeks ago, we outfitted Chloe and Erik for the their first “official” bikepacking trip - a weekend of camping and riding north from San Francisco to Wildcat Campground, on the Coast Trail in Bolinas. As with any “firsts” like these, there were bound to be some lessons learned along with many pleasant surprises. Chloe wrote the bits below, and they sent along some lovely photos to go along with it. Have a read and hear how it went for these two!</p>
         <p>When your friends snag a last minute campsite at Wildcat, you jump on the offer, no questions asked.  That’s how Erik and I ended up scrambling to prep our newly acquired gravel bikes for our first ~proper~ bikepacking trip. The plan was simple: ride up to Olema from San Francisco on Saturday morning (about 45 miles), hit the Five Brooks trailhead early in the afternoon, and arrive at the coastal campsite an hour or so later.  Since we would only be camping one night and would also be meeting friends who were hiking in, we were able to pack light.  Even so, it took several re-orgs to fit all the essentials: clothes and sandals in the Outer Shell Handlebar bag, tools and food in the Framebag, and sleeping bag, pad, and tent in the Expedition Seatbag.  Once in place, however, the rigs were surprisingly light, maneuverable, and OF COURSE  stylish (camo vs. marigold--who wore it better?!).</p>
         <img src="https://rstr.in/monogram/vibes/szWSuA7-jTp" alt="Two people standing next to their bikes, both wearing helmets. They are smiling and looking at the camera." />
@@ -20,11 +93,30 @@ export default function Preview() {
 
         <p>Before I even went to bed that night, I had decided that this was one of my favorite camping trips ever.  The  combination of surreal scenery, wonderful crew, and adrenaline of the ride made it hard to beat.  I was challenged by both the total distance and some of the terrain, but never felt in over my head.  I got a better feel for my bike, learned how to pack efficiently, and improved my navigation skills.  Best of all, it was so easy to do! In just one week we were able to assemble all the bags and camping gear needed to make it happen. There was no car involved.  We could go anywhere, road or dirt, carrying everything we needed. It’s hard to feel more free than that.</p>
       `}
-      relatedPostsTitle="Related posts"
-      cta={{
-        href: '#',
-        label: 'View All',
-      }}
-    />
+      />
+
+      <FeaturedBlogPostCarousel
+        title="Related posts"
+        blogPosts={posts}
+        cta={{
+          href: '#',
+          label: 'View All',
+        }}
+      />
+
+      <Subscribe
+        title="Sign up for our newsletter"
+        description="Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor."
+      />
+
+      <Footer
+        logo={logo}
+        sections={footerLinks}
+        contactInformation={contactInformation}
+        copyright={copyright}
+        paymentIcons={paymentIconsArray}
+        socialMediaLinks={socialMediaLinks}
+      />
+    </>
   )
 }
